+++ src/dotnet/ReSharperPlugin.SpecflowRiderPlugin/Daemon/SpecflowHighlightingBase.cs
@@ -1,0 +1,8 @@
+namespace ReSharperPlugin.SpecflowRiderPlugin.Daemon
+{
+    public abstract class SpecflowHighlightingBase
+    {
+        // ErrorsGen makes IsValid override if we specify a base class
+        public abstract bool IsValid();
+    }
+}