+++ src/dotnet/ReSharperPlugin.SpecflowRiderPlugin/QuickFixes/MethodNameMismatchPatternQuickFix.cs
@@ -1,0 +1,41 @@
+using System;
+using JetBrains.Application.Progress;
+using JetBrains.ProjectModel;
+using JetBrains.ReSharper.Feature.Services.Bulbs;
+using JetBrains.ReSharper.Feature.Services.QuickFixes;
+using JetBrains.ReSharper.Feature.Services.Refactorings.Specific.Rename;
+using JetBrains.ReSharper.Intentions.Xaml.QuickFixes;
+using JetBrains.ReSharper.Plugins.Unity.CSharp.Daemon.Errors;
+using JetBrains.ReSharper.Psi.CSharp;
+using JetBrains.ReSharper.Psi.CSharp.Impl;
+using JetBrains.ReSharper.Refactorings.CSharp;
+using JetBrains.ReSharper.Refactorings.CSharp.Rename;
+using JetBrains.TextControl;
+using JetBrains.Util;
+
+namespace ReSharperPlugin.SpecflowRiderPlugin.QuickFixes
+{
+    [QuickFix]
+    public class MethodNameMismatchPatternQuickFix : QuickFixBase
+    {
+        private readonly MethodNameMismatchPatternWarning _warning;
+
+        public override string Text => "Rename to " + _warning.ExpectedName;
+
+        public MethodNameMismatchPatternQuickFix(MethodNameMismatchPatternWarning warning)
+        {
+            _warning = warning;
+        }
+
+        protected override Action<ITextControl> ExecutePsiTransaction(ISolution solution, IProgressIndicator progress)
+        {
+            CSharpImplUtil.ReplaceIdentifier(_warning.Method.NameIdentifier, _warning.ExpectedName);
+            return _ => { };
+        }
+
+        public override bool IsAvailable(IUserDataHolder cache)
+        {
+            return true;
+        }
+    }
+}